requirements.txt
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
playingcards.py==1.1.1
pluggy==1.5.0
pyfiglet==1.0.2
Pygments==2.18.0
pytest==8.2.0
rich==13.7.1
